+++ packages/agent-infra/browser-use/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @agent-infra/browser-use
 
+## 0.1.1-beta.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [bd34a4e]
+  - @agent-infra/shared@0.0.2-beta.0
+  - @agent-infra/browser@0.1.1-beta.2
+
+## 0.1.1-beta.1
+
+### Patch Changes
+
+- chore: publish
+- Updated dependencies
+  - @agent-infra/browser@0.1.1-beta.1
+  - @agent-infra/shared@0.0.1
+
+## 0.1.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @agent-infra/browser@0.1.1-beta.0
+  - @agent-infra/shared@0.0.1
+
 ## 0.1.0
 
 ### Minor Changes