+++ packages/agent-infra/mcp-client/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @agent-infra/mcp-client
 
+## 1.1.6-beta.4
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.4
+
+## 1.1.6-beta.3
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.3
+
+## 1.1.6-beta.2
+
+### Patch Changes
+
+- chore: publish
+- Updated dependencies
+  - @agent-infra/mcp-shared@1.1.6-beta.2
+
+## 1.1.6-beta.1
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.1
+
+## 1.1.6-beta.0
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.0
+
+## 1.1.5
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.4
+
+## 1.1.3
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.3
+
+## 1.1.2
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes