+++ packages/agent-infra/mcp-servers/commands/CHANGELOG.md
@@ -1,5 +1,47 @@
 ## 0.4.0 2024-12-10 - add logging
 
+## 1.1.6-beta.4
+
+## 1.1.6-beta.3
+
+## 1.1.6-beta.2
+
+### Patch Changes
+
+- chore: publish
+
+## 1.1.6-beta.1
+
+## 1.1.6-beta.0
+
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.1.5
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.1.4
+
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.1.3
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes