+++ packages/agent-infra/mcp-shared/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @agent-infra/mcp-shared
 
+## 1.1.6-beta.4
+
+## 1.1.6-beta.3
+
+## 1.1.6-beta.2
+
+### Patch Changes
+
+- chore: publish
+
+## 1.1.6-beta.1
+
+## 1.1.6-beta.0
+
+## 1.1.5
+
+## 1.1.4
+
+## 1.1.3
+
+## 1.1.2
+
 ## 1.1.1
 
 ### Patch Changes