+++ packages/agent-infra/search/search/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @agent-infra/search
 
+## 0.0.3-beta.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [bd34a4e]
+  - @agent-infra/shared@0.0.2-beta.0
+  - @agent-infra/browser-search@0.0.3-beta.3
+
+## 0.0.3-beta.2
+
+### Patch Changes
+
+- chore: publish
+- Updated dependencies
+  - @agent-infra/browser-search@0.0.3-beta.2
+  - @agent-infra/shared@0.0.1
+
+## 0.0.3-beta.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @agent-infra/browser-search@0.0.3-beta.1
+  - @agent-infra/shared@0.0.1
+
+## 0.0.3-beta.0
+
+### Patch Changes
+
+- @agent-infra/browser-search@0.0.3-beta.0
+- @agent-infra/shared@0.0.1
+
 ## 0.0.2
 
 ### Patch Changes